fix: add game search to shop view models
