+++ app/src/main/java/com/github/meeplemeet/model/shops/CreateShopViewModel.kt
@@ -2,7 +2,6 @@
 
 package com.github.meeplemeet.model.shops
 
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.github.meeplemeet.RepositoryProvider
 import com.github.meeplemeet.model.auth.Account
@@ -16,10 +15,10 @@ import kotlinx.coroutines.launch
  * This ViewModel handles the creation of board game shops with validation logic to ensure all
  * required fields are properly provided.
  *
- * @property repository The repository used for shop operations.
+ * @property shopRepo The repository used for shop operations.
  */
-class CreateShopViewModel(private val repository: ShopRepository = RepositoryProvider.shops) :
-    ViewModel() {
+class CreateShopViewModel(private val shopRepo: ShopRepository = RepositoryProvider.shops) :
+    ShopGameViewModel() {
   /**
    * Creates a new shop in Firestore.
    *
@@ -59,8 +58,7 @@ class CreateShopViewModel(private val repository: ShopRepository = RepositoryPro
         throw IllegalArgumentException("An address it required to create a shop")
 
     viewModelScope.launch {
-      repository.createShop(
-          owner, name, phone, email, website, address, openingHours, gameCollection)
+      shopRepo.createShop(owner, name, phone, email, website, address, openingHours, gameCollection)
     }
   }
 }
