+++ app/src/main/java/com/github/meeplemeet/model/shops/EditShopViewModel.kt
@@ -2,7 +2,6 @@
 
 package com.github.meeplemeet.model.shops
 
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.github.meeplemeet.RepositoryProvider
 import com.github.meeplemeet.model.PermissionDeniedException
@@ -20,7 +19,7 @@ import kotlinx.coroutines.launch
  * @property repository The repository used for shop operations.
  */
 class EditShopViewModel(private val repository: ShopRepository = RepositoryProvider.shops) :
-    ViewModel() {
+    ShopGameViewModel() {
 
   /**
    * Updates one or more fields of an existing shop.
