+++ app/src/main/java/com/github/meeplemeet/model/shops/ShopViewModel.kt
@@ -2,7 +2,6 @@
 
 package com.github.meeplemeet.model.shops
 
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.github.meeplemeet.RepositoryProvider
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -17,7 +16,7 @@ import kotlinx.coroutines.launch
  * @property repository The repository used for shop operations.
  */
 class ShopViewModel(private val repository: ShopRepository = RepositoryProvider.shops) :
-    ViewModel() {
+    ShopGameViewModel() {
   private val _shop = MutableStateFlow<Shop?>(null)
 
   /**
